+++ src/util.h
@@ -15,11 +15,11 @@ static inline double reldist(numeric x, numeric y, double eps)
 }
 
 
-double uniform_distance_real(size_t n, const double* x, const double* y);
+double uniform_distance_real(long n, const double* x, const double* y);
 
-double uniform_distance(size_t n, const numeric* x, const numeric* y);
+double uniform_distance(long n, const numeric* x, const numeric* y);
 
-double relative_distance(size_t n, const numeric* x, const numeric* y, double eps);
+double relative_distance(long n, const numeric* x, const numeric* y, double eps);
 
 
 herr_t read_hdf5_dataset(hid_t file, const char* name, hid_t mem_type, void* data);
